Migrate instance-ID APIs to EntityId (6000.3+), widen handles to long (#55)
* Fix Unity 6000.5 compatibility: use EntityId APIs

Unity 6000.5 renamed the int-based instance ID APIs to the 64-bit
EntityId type and marked Object.GetInstanceID() and
EditorUtility.InstanceIDToObject(int) obsolete-as-error (CS0619).

Add version-gated wrappers (GetObjectInstanceId / ResolveObjectById)
that use the new EntityId APIs on Unity 6000.5+ and fall back to the
legacy APIs on older versions (package still targets Unity 2022.3+).
Instance-ID handles remain int across the protocol; on 6000.5+ they
round-trip through EntityId.ToULong()/FromULong().

Also route the CLI's generated eval --id preamble through the new
ResolveObjectById helper so it compiles under Unity 6000.5.

* @
Widen instance-id handles to long; adopt EntityId from 6000.3

Builds on the 6000.5 EntityId migration. Moves the #if gate to 6000.3,
where the int-based instance-ID APIs are already deprecated (CS0618) and
EntityId exists, so 6.3/6.4 stop emitting deprecation warnings. Uses a
three-way path because EntityId is int-backed on 6000.3/.4 (implicit
int conversions) and only 64-bit (ToULong/FromULong) on 6000.5+.

Surfaces handles as long end-to-end so a 6000.5 64-bit EntityId is never
truncated; narrower ids widen and narrow back exactly:
- Protocol DTOs: InstanceId, OpenedFromInstanceId, BlockedBy
- CLI: --id options (snapshot, ui click), ParseInstanceIds + eval preamble
- Package: wrappers, GetLongArgument, targetId/context paths, StageInfo
- Rebuilt bundled UnityCtl.Protocol.dll

Verified on 6000.3.15f1: compiles clean (no Safe Mode), deprecation
warnings gone, snapshot --id round-trips. dotnet build + 482 tests pass.
@

---------

+++ UnityCtl.Cli/UICommands.cs
@@ -20,7 +20,7 @@ public static class UICommands
     {
         var clickCommand = new Command("click", "Click a UI element by instance ID, name, or screen coordinates");
 
-        var idOption = new Option<int?>("--id", "Instance ID of the UI element to click");
+        var idOption = new Option<long?>("--id", "Instance ID of the UI element to click");
         var nameOption = new Option<string?>("--name", "Find and click a GameObject by name (uses GameObject.Find)");
         var xArg = new Argument<int?>("x", () => null, "Screen X coordinate");
         var yArg = new Argument<int?>("y", () => null, "Screen Y coordinate");
